Move DerivedSignal to derived.py from __init__.py
